fix(battery): Colors / icons not updating
I didn't understand that theme.sh is only run once when configuring the
tmux status line. As a result, the colors and and icon for the battery
plugin were static and didn't update based on state as it changed.

To fix that, I had to set some placeholders and pass the whole section
of the battery plugin as a template to the plugin itself so it could set
the icon/colors dynamically, based on the state. It's not the cleanest
but I didn't want to have to duplicate all of the separator / ordering
logic so felt like this was the best approach.